pyproject.toml
[project]
name = "t3-receitasapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.1",
    "fastapi>=0.128.0",
    "fastapi-pagination>=0.15.6",
    "motor>=3.7.1",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.9.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

